Renamed logging to logger
- Avoid's conflict with standard logging module
