+++ docs/logger.rst
@@ -1,0 +1,7 @@
+Logger
+=======
+
+Coaster can help your application log errors at run-time. Initialize with :func:`coaster.logger.init_app`.
+
+.. automodule:: coaster.logger
+   :members:
