--- docs/logging.rst
@@ -1,7 +1,0 @@
-Logging
-=======
-
-Coaster can help your application log errors at run-time. Initialize with :func:`coaster.logging.init_app`.
-
-.. automodule:: coaster.logging
-   :members:
